pyproject.toml
[project]
name = "testtask"
version = "0.1.0"
description = ""
authors = [
    {name = "Namra404",email = "armeno123@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "fastapi (>=0.121.2,<0.122.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false